Refactor key handling functionality
