+++ eval.go
@@ -1713,7 +1713,7 @@ func (e *callExpr) eval(app *app, _ []string) {
 		}
 		log.Println("pushing keys", e.args[0])
 		for _, val := range splitKeys(e.args[0]) {
-			app.ui.keyChan <- val
+			app.ui.evChan <- parseKey(val)
 		}
 	case "addcustominfo":
 		var k, v string
